+++ CHANGELOG.md
@@ -7,5 +7,6 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added
 
+  - 2016-11-07 Added `String/ConversionTrait.php`
   - 2016-10-01 Added `enableDeRefence()` in `ReferenceInterface`
   - 2016-09-30 Added `getLanguage()` in `LoaderInterface`